refactor: move all server files to server folder
- install serverless-http module to wrap the express app
- commented legacy code in server.js
- for all the routes that are not specified, return 404
