+++ server/package.json
@@ -41,7 +41,8 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^6.0.14"
+    "mongoose": "^6.0.14",
+    "serverless-http": "^3.0.1"
   },
   "config": {
     "commitizen": {
